# [REF] base: Speed-up attachments searches bypassing "unaccent" method call