--- a/odoo/addons/base/models/ir_attachment.py
+++ b/odoo/addons/base/models/ir_attachment.py
@@ -400,7 +400,7 @@ class IrAttachment(models.Model):
     type = fields.Selection([('url', 'URL'), ('binary', 'File')],
                             string='Type', required=True, default='binary', change_default=True,
                             help="You can either upload a file from your computer or copy/paste an internet link to your file.")
-    url = fields.Char('Url', index='btree_not_null', size=1024)
+    url = fields.Char('Url', index='btree_not_null', size=1024, unaccent=False)
     public = fields.Boolean('Is public document')
 
     # for external access