@robodoo 2023-06-12
[Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/124761).