@HydrionBurst 2023-07-27
I think we should check the reason for index failure 🤔 
Could you check if your index function contains `unaccent` in your use case with `unaccent`?